create components folder, remove posts from homepage
