+++ src/lib/CreatePostForm.svelte
@@ -1,15 +1,28 @@
+<script>
+  import Button from "./components/Button.svelte";
+</script>
+
 <form action="?/createPost" method="POST">
   <h2>What do you want to say?</h2>
   <label for="title"><h3>Title</h3></label>
   <input class="formEntry" type="text" id="title" name="title" />
   <label for="content"><h3>Content</h3></label>
   <textarea class="formEntry" id="content" name="content" rows={5} />
-  <button type="submit">Add Post</button>
+  <div class="buttonContainer">
+    <Button type="submit" text="Add Post" />
+  </div>
 </form>
 
 <style>
   form {
-    display: grid;
+    display: block;
+
+    border: solid gray;
+    border-radius: 10px;
+
+    width: fit-content;
+
+    padding: 1%;
   }
 
   .formEntry {
@@ -19,10 +32,5 @@
     width: fit-content;
     border: solid gray;
     border-radius: 10px;
-  }
-
-  button {
-    margin-top: 5%;
-    width: fit-content;
   }
 </style>
