+++ src/lib/components/Button.svelte
@@ -1,0 +1,15 @@
+<script>
+  export let type;
+  export let text;
+</script>
+
+<button {type}>{text}</button>
+
+<style>
+  button {
+    margin: 0 auto;
+    width: fit-content;
+    background-color: lightblue;
+    border-radius: 10px;
+  }
+</style>
