+++ src/lib/components/ClickableCard.svelte
