+++ src/routes/onboard/+page.svelte
@@ -6,6 +6,9 @@
 <Navbar />
 <main>
   <h1>Onboarding</h1>
+
   <h2>Here are some users. Are any of them in your social circle?</h2>
   <Grid />
+
+  <h2>FYI: This page is WIP. These are fake users.</h2>
 </main>
