+++ src/routes/origin/+page.server.ts
@@ -29,4 +29,22 @@ export const actions: Actions = {
 
     return { status: 201 };
   },
+
+  deletePost: async ({ url }) => {
+    console.log("deletePost called");
+    const id = url.searchParams.get("id");
+    if (!id) {
+      return fail(500, { message: "invalid request" });
+    }
+    try {
+      await prisma.spheres_posts.delete({ where: { id: Number(id) } });
+    } catch (err) {
+      console.error(err);
+      return fail(500, {
+        message: "Something went wrong deleting your article",
+      });
+    }
+
+    return { status: 200 };
+  },
 };
